fix(dashboard): guard accumulateEvent against missing ide field
Extract the shared createAccumulator() factory into dashboard/accumulator.js
so the test exercises the real production code path instead of a stale
inline copy.

Changes:
- accumulator.js: new module with createAccumulator(prices) factory
- server.js: import accumulator, remove 97 lines of inline definitions
- tests/dashboard-server.test.js: import accumulator, remove 115-line
  inline copy; all 8 assertions unchanged

The fix itself:
Add early-return guard clause in accumulateEvent() to prevent
getProvider(undefined) from creating a phantom provider entry
under the key 'undefined' when a telemetry event has no 'ide'
field. This corrupted /telemetry and /cost-summary views.

Broadcast guard: accumulateEvent now returns true/false so the
POST /event handler only broadcasts accepted events to WebSocket
clients, preventing malformed events from reaching subscribers.

Regression test added covering normal events, missing 'ide',
null/undefined input, empty string validation, and return-value
contract for broadcast gating.

Fixes #500
