+++ dashboard/server.js
@@ -6,6 +6,7 @@ const path    = require('path');
 const fs      = require('fs');
 const os      = require('os');
 const { execSync } = require('child_process');
+const { createAccumulator } = require('./accumulator');
 
 const PORT   = 7890;
 const PUBLIC = path.join(__dirname, 'public');
@@ -46,119 +47,27 @@ function detectOperator() {
 }
 const OPERATOR = detectOperator();
 
-// What each provider can actually expose — honest contract
-const CAPABILITIES = {
-  claude:    { tokenUsage:true,  model:true,  cost:true,  session:true,  workspace:true  },
-  gemini:    { tokenUsage:true,  model:true,  cost:false, session:true,  workspace:false },
-  cursor:    { tokenUsage:false, model:false, cost:false, session:true,  workspace:true  },
-  codex:     { tokenUsage:false, model:false, cost:false, session:false, workspace:false },
-  vscode:    { tokenUsage:false, model:false, cost:false, session:false, workspace:true  },
-  kiro:      { tokenUsage:false, model:false, cost:false, session:false, workspace:false },
-  trae:      { tokenUsage:false, model:false, cost:false, session:false, workspace:false },
-  opencode:  { tokenUsage:false, model:false, cost:false, session:false, workspace:false },
-  codebuddy: { tokenUsage:false, model:false, cost:false, session:false, workspace:false },
-  aider:     { tokenUsage:false, model:false, cost:false, session:false, workspace:false },
-};
-
 // Pricing per 1M tokens — loaded from prices.json (configurable)
 const PRICES_PATH = path.join(__dirname, 'prices.json');
-let MODEL_PRICES = {};
+const MODEL_PRICES = {};
 function loadPrices() {
   try {
-    MODEL_PRICES = JSON.parse(fs.readFileSync(PRICES_PATH, 'utf8'));
+    const data = JSON.parse(fs.readFileSync(PRICES_PATH, 'utf8'));
+    Object.assign(MODEL_PRICES, data);
   } catch (_) {
-    MODEL_PRICES = {
+    Object.assign(MODEL_PRICES, {
       '_default_claude': { input: 3.00, output: 15.00, cacheRead: 0.30, cacheWrite: 3.75 },
       '_default_gemini': { input: 0.10, output: 0.40,  cacheRead: 0.025, cacheWrite: 0.00 },
       '_default_codex':  { input: 2.50, output: 10.00, cacheRead: 1.25, cacheWrite: 0.00 },
-    };
+    });
   }
 }
 loadPrices();
 fs.watchFile(PRICES_PATH, () => loadPrices());
 
-const IDE_PRICE_KEY = {
-  claude: '_default_claude',
-  gemini: '_default_gemini',
-  codex:  '_default_codex',
-  opencode: '_default_opencode',
-};
-
-function calcCost(ide, tokens, model) {
-  const pricing = MODEL_PRICES[model] || MODEL_PRICES[IDE_PRICE_KEY[ide]];
-  if (!pricing) return null;
-  const inp = (tokens.input      || 0) * (pricing.input      || 0) / 1e6;
-  const out = (tokens.output     || 0) * (pricing.output     || 0) / 1e6;
-  const cr  = (tokens.cacheRead  || 0) * (pricing.cacheRead  || 0) / 1e6;
-  const cw  = (tokens.cacheWrite || 0) * (pricing.cacheWrite || 0) / 1e6;
-  return inp + out + cr + cw;
-}
-
-// Runtime telemetry accumulated from real events
-const providerState = {};
-// Store recent completed sessions for dashboard analytics
-const sessionHistory = [];
-const MAX_SESSION_HISTORY = 1000;
-
-function getProvider(ide) {
-  if (!providerState[ide]) {
-    providerState[ide] = {
-      ide,
-      lastSeen:  null,
-      running:   false,
-      toolCalls: 0,
-      sessions:  0,
-      lastModel: null,          // ← add this line
-      tokens: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0 },
-    };
-  }
-  return providerState[ide];
-}
-
-function accumulateEvent(ev) {
-  
-
-  const p = getProvider(ev.ide);
-  p.lastSeen = Date.now();
-  p.running  = true;
-  if (ev.model) p.lastModel = ev.model;   // ← add this line
-
-if (ev.event === 'pre_tool') p.toolCalls++;
-
-if (ev.event === 'session_end') {
-  const usage = ev.usage || {};
-  const sessionTokens = {
-    input:      usage.input_tokens                || 0,
-    output:     usage.output_tokens               || 0,
-    cacheRead:  usage.cache_read_input_tokens     || 0,
-    cacheWrite: usage.cache_creation_input_tokens || 0,
-  };
-  const sessionModel = ev.model || p.lastModel || null;
-  const ideCap       = CAPABILITIES[ev.ide] || {};
-  const sessionCost  = ideCap.cost === true ? calcCost(ev.ide, sessionTokens, sessionModel) : null;
-
-  sessionHistory.push({
-    timestamp:    Date.now(),
-    ide:          ev.ide,
-    model:        sessionModel,
-    input_tokens:  sessionTokens.input,
-    output_tokens: sessionTokens.output,
-    total_tokens:  sessionTokens.input + sessionTokens.output,
-    cost:          sessionCost,
-  });
-
-  if (sessionHistory.length > MAX_SESSION_HISTORY) sessionHistory.shift();
-}
-
-if (ev.usage) {
-  p.sessions++;
-
-  p.tokens.input += (ev.usage.input_tokens || 0);
-  p.tokens.output += (ev.usage.output_tokens || 0);
-  p.tokens.cacheRead += (ev.usage.cache_read_input_tokens || 0);
-  p.tokens.cacheWrite += (ev.usage.cache_creation_input_tokens || 0);
-}
-}
+// Shared accumulator — fresh state, production logic
+const ACC = createAccumulator(MODEL_PRICES);
+const { providerState, sessionHistory, getProvider, accumulateEvent, calcCost, CAPABILITIES } = ACC;
 
 // Mark providers offline after 90 s without events
 setInterval(() => {
@@ -189,9 +98,10 @@ const server = http.createServer((req, res) => {
     req.on('end', () => {
       try {
         const ev = JSON.parse(body);
-        accumulateEvent(ev);
-        const msg = JSON.stringify(ev);
-        for (const ws of clients) { if (ws.readyState === 1) ws.send(msg); }
+        if (accumulateEvent(ev)) {
+          const msg = JSON.stringify(ev);
+          for (const ws of clients) { if (ws.readyState === 1) ws.send(msg); }
+        }
       } catch (_) {}
       res.writeHead(200, { 'Content-Type': 'application/json' });
       res.end('{"ok":true}');
